diary java commit

+++ minipage/src/main/java/com/mysite/minipage/diary/DiaryController.java
@@ -6,6 +6,8 @@ import java.util.List;
 import javax.validation.Valid;
 
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.server.ResponseStatusException;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,21 +28,23 @@ public class DiaryController {
 	private final DiaryService diaryService;
 	
 	@RequestMapping("/list")
-	public String list(Model model) {
-		List<Diary> diaryList = this.diaryService.getList();
+	public String list(Model model, @RequestParam(value = "page", defaultValue = "0") int page) {
+		Page<Diary> paging = this.diaryService.getList(page);
 		
-		model.addAttribute("diaryList", diaryList);
+		model.addAttribute("paging", paging);
 		
 		return "diary";
 	}
 	
 	/* 글 등록 */
-	@GetMapping("/create")
+	@PreAuthorize("isAuthenticated()")
+	@GetMapping("/list/create")
 	public String diaryCreate(DiaryForm diaryForm) {
 		return "diary_form";
 	}
 	
-	@PostMapping("/create")
+	@PreAuthorize("isAuthenticated()")
+	@PostMapping("/list/create")
 	public String diaryCreate(@Valid DiaryForm diaryForm, BindingResult bindingResult) {
 		if(bindingResult.hasErrors()) {
 			return "diary_form";
@@ -50,24 +55,32 @@ public class DiaryController {
 	}
 	
 	/* 글 수정 */
+	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/modify/{id}")
-	public String diaryModify(DiaryForm diaryForm, @PathVariable("id") Integer id) {
-		 Diary diary = this.diaryService.getDiary(id);
-		 
+	public String diaryModify(DiaryForm diaryForm, @PathVariable("id") Integer id, Principal principal) {
+		Diary diary = this.diaryService.getDiary(id);
+		if(!diary.getAuthor().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+		}
+		
 		diaryForm.setContent(diary.getContent());
 		
 		return "diary_form";
 	 }
 	
+	@PreAuthorize("isAuthenticated()")
 	@PostMapping("/modify/{id}")
 	public String diaryModify(@Valid DiaryForm diaryForm, BindingResult bindingResult,
-			@PathVariable("id") Integer id)
+			Principal principal, @PathVariable("id") Integer id)
 	{
 		if(bindingResult.hasErrors()) {
 			return "diary_form";
 		}
 		
 		Diary diary = this.diaryService.getDiary(id);
+		if(!diary.getAuthor().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+		}
 		
 		this.diaryService.modify(diary, diaryForm.getContent());
 		
@@ -75,9 +88,13 @@ public class DiaryController {
 	}
 	
 	/* 글 삭제 */
+	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/delete/{id}")
 	public String diaryDelete(Principal principal, @PathVariable("id") Integer id) {
 		Diary diary = this.diaryService.getDiary(id);
+		if(!diary.getAuthor().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+		}
 		
 		this.diaryService.delete(diary);
